App/requirements.txt
telethon
pytz
pymongo
python-dotenv